Fixed warnings.
git-svn-id: https://svn.aros.org/svn/aros/trunk@52197 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ tools/flexcat/src/createcatsrc.c
@@ -666,7 +666,7 @@ void CreateSourceFile(char *SourceFile, char *TemplateFile, char *CDFile)
                     char *start;
                     char _StrLen[20 + 1];
 
-                    snprintf(_StrLen, sizeof(_StrLen), "%020lx", (uint32)cs->ID);
+                    snprintf(_StrLen, sizeof(_StrLen), "%020lx", (long unsigned)cs->ID);
                     start = &_StrLen[20 - _len * 2];
                     while(_len > 0)
                     {
@@ -681,7 +681,7 @@ void CreateSourceFile(char *SourceFile, char *TemplateFile, char *CDFile)
                     char *start;
                     char _StrLen[20 + 1];
 
-                    snprintf(_StrLen, sizeof(_StrLen), "%020lx", (uint32)((CalcRealLength(cs->CD_Str) + 1) & 0xfffffe));
+                    snprintf(_StrLen, sizeof(_StrLen), "%020lx", (long unsigned)((CalcRealLength(cs->CD_Str) + 1) & 0xfffffe));
                     start = &_StrLen[20 - _len * 2];
                     while(_len > 0)
                     {
